Simple querying
At this point the good-path works, as long as the result fits in memory.
You still can't upsert though -- the only way to get data in is to create
a dataset with soda-fountain, then publish-upsert-secondarize by talking
straight to the data-coordinator.  Once it's there in the secondary, though,
it works.

Errors aren't translated.