Add a "_" prefix when we map column names to ids for rollup tables to prevent invalid identifiers.
When we pass the soql query defining a rollup table along to the data coordinator, we want to map
column names to something the lower layers of the stack knows about.  However, unfortunately some
of our internal column ids start with numbers which is not valid in soql.  To avoid this problem,
we prefix any non-system columns with a "_".