Allow single-row upsert
This is a stupid misfeature of soda2, but it's something the fakesoda2
supports and for some reason it's wanted in realsoda2 too.  This just
makes the upsert of a single object the same as upserting a singleton
list containing that item, by detecting whether the first event of the
input is a StartOfArrayEvent, and if not producing a singleton iterator
of the result of reading a value.