Refs CORE-5975
Add a GET request query parameter to the /export endpoint
which will explicitly specify the CRS of the export.
Not going to name names, but this is needed
for certain libraries that crash when you parse geoJSON
which follows the spec correctly and includes the CRS.
Currently only crs84 is supported, and that is all that
will be supported because anything else will break the
geoJSON spec.