Multiple adjustment to errors and their respective classes. The idea is to handle the complete set of errors we can get from Data-Coordinator, all the changes should reflect this. In addition made a few adjustments to the tests but these are minor.
updated cleaned and trimmed the export and column DAOs.

kept the expected row DAO results, and cleaned up.

done with dataset DAOs. Next step is to check agains apis.

updated suggestTest with correct error and created a new test for the HttpDataCoordinatorClient.

multiple changes to fit query coordinator errors in some fashion similar to data coordinator. However, it seems that sf may only take in the errors and punt them up as a sf error and that is all. So this means that we may not need the same level of granularity. Just in case this may not be true in the future I have added the machinery to handle the situation as it arises in the future.

refactored HttpQueryCoordinatorClient in order to create a test to ensure that we get the type of errors we expect to retrieve from query-coordinator.

surface dc's command index in soda-fountain; add in one useful error log message to show how/where to do that; differentiate between row-has-null-in-pk-column and row-not-found

add another readable message; actually use the human-readable thing in the 'responding with error' log message

basic error message support; switching to using the sanitized data for this, though probably there are still some things we want to override to not show additional stacktraces, etc by default. tested versus computed columns errors and seems to provide saner information, although there is the potential to leak things we don't want (one error will complain if you try a GET and it only supports a POST -- and it will tell you the things it supports), so should look through the info these things have more carefully before releasing this

updated Resource errors, cleaned HttpQueryCoordinator a bit and added status to internalservice error for complete metric tracking.

cleaned up Resource.scala and left a note for a method called response which doesnt seem to be required.

renamed an error class, and cleaned up the Dataset resource object.

few changes to sf to capture additional dataset errors that can occur, this should complete the missing error and content as left in todos.

slight clean up for dataset.scala

rename some errors to make actual sense and reflect what they actually mean; add pretty messages to errors instead of just having tags

pull in query coordinator error descriptions and use those instead of just outright printing the data values as part of the message (they get surfaced alongside the message where relevant)

fix tests for qc/dc to match the current expected error formats

made error catching for unknown errors better and more detailed.