Stop logging about not getting a user passed in, we know we don't do that.
We don't pass user infomation down to sf so it just logs a silly
message.   Not ripping it out entirely because that would also require changes
in DC... we certainly could easily start passing down a user for debugging info but that
wouldn't be in the Authorization header since it isn't auth.