EN-38315: Don't try to snapshot deleted datasets.
Currently snapshotter is looping constantly trying to snapshot
datasets that are deleted, and failing because they can't be accessed
because they are marked deleted.  To avoid this we filter out deleted datasets.

The only other uses of bulkDatasetLookup are in collation related code, and they
don't make sense for deleted datasets either and/or already won't work if the dataset
is deleted.  I'm adding it as a paramter instead of hardcoding it to help any future
users of the method think about if they want to include deleted datasets or not.