Remove "spandex connect timeout" test.
There is no guaranteed platform independant way to ensure you will get a connection
timeout instead of a faster failure without specifically setting up an IP that generates
no response at all to connection attempts or doing platfom specific configuration on the
host the tests are running on (eg. firewall rules).

In my specific case, connections to private IP addresses get an ICMP "admin prohibited"
response which turns into an immediate connection refused, which is a valid way for networks
to handle outbound traffic to RFC-1918 address space.

Given that, the value of this test and what it is actually testing is low enough that removing
it seems to make the most sense.