Allow doing soql queries via QUERY rather than GET
Everything below this level is already done via not-GET (specifically,
done via POST, but POST already means something for accessing a
`/resource/` endpoint).

The format of the QUERY endpoint is the same as for GET, but rather
than taking parameters on the URL, it receives a body containing
a JSON object of strings which are treated in the same way.